tbe-3: updated EducationPairsHelperTelegramBot class
