+++ src/main/java/com/edication/telegram/bot/telegramboteducationlecture/bot/EducationPairsHelperTelegramBot.java
@@ -1,0 +1,65 @@
+package com.edication.telegram.bot.telegramboteducationlecture.bot;
+
+import com.edication.telegram.bot.telegramboteducationlecture.command.CommandContainer;
+import com.edication.telegram.bot.telegramboteducationlecture.service.SendBotMessageServiceImpl;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+import org.telegram.telegrambots.bots.TelegramLongPollingBot;
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
+import org.telegram.telegrambots.meta.api.objects.Update;
+import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
+
+import java.util.Locale;
+
+import static com.edication.telegram.bot.telegramboteducationlecture.command.CommandName.NO;
+
+@Component
+public class EducationPairsHelperTelegramBot extends TelegramLongPollingBot {
+
+
+    private static final String COMMAND_PREFIX = "/";
+    @Value("${bot.name}")
+    private String botName;
+    @Value("${bot.token}")
+    private String botToken;
+
+    private final CommandContainer commandContainer;
+
+    public EducationPairsHelperTelegramBot() {
+        this.commandContainer = new CommandContainer(new SendBotMessageServiceImpl(this));
+    }
+
+    @Override
+    public String getBotUsername() {
+        return botName;
+    }
+
+    @Override
+    public String getBotToken() {
+        return botToken;
+    }
+
+    @Override
+    public void onUpdateReceived(Update update) {
+        if (update.hasMessage() && update.getMessage().hasText()){
+            String message = update.getMessage().getText().trim();
+            if (message.toLowerCase().startsWith("леша") || message.toLowerCase().startsWith("лёша")){
+                SendMessage sendMessage = new SendMessage();
+                sendMessage.setChatId(update.getMessage().getChatId().toString());
+                sendMessage.setText("Та внатуре.\nПодтверждаю окоянный)");
+
+                try {
+                    execute(sendMessage);
+                } catch (TelegramApiException e) {
+                    e.printStackTrace();
+                }
+            }
+            else if (message.startsWith(COMMAND_PREFIX)){
+                String commandIdentifier = message.split(" ")[0].toLowerCase();
+                commandContainer.retrieveCommand(commandIdentifier).execute(update);
+            }else {
+                commandContainer.retrieveCommand(NO.getCommandName()).execute(update);
+            }
+        }
+    }
+}
