--- src/main/java/com/edication/telegram/bot/telegramboteducationlecture/components/EducationPairsHelperTelegramBot.java
@@ -1,52 +1,0 @@
-package com.edication.telegram.bot.telegramboteducationlecture.components;
-
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.stereotype.Component;
-import org.telegram.telegrambots.bots.TelegramLongPollingBot;
-import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
-import org.telegram.telegrambots.meta.api.objects.Update;
-import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
-
-@Component
-public class EducationPairsHelperTelegramBot extends TelegramLongPollingBot {
-
-
-
-    @Value("${bot.name}")
-    private String botName;
-    @Value("${bot.token}")
-    private String botToken;
-
-
-
-    @Override
-    public String getBotUsername() {
-        return botName;
-    }
-
-    @Override
-    public String getBotToken() {
-        return botToken;
-    }
-
-    @Override
-    public void onUpdateReceived(Update update) {
-        if (update.hasMessage() && update.getMessage().hasText()){
-            String message = update.getMessage().getText().trim();
-            String chatId  = update.getMessage().getChatId().toString();
-
-            SendMessage sm  = new SendMessage();
-            sm.setChatId(chatId);
-            sm.setText(message);
-
-            try {
-                execute(sm);
-            } catch (TelegramApiException e) {
-                e.printStackTrace();
-            }
-        }
-    }
-}
